Make links explicit in text-only version
This relates to how Assistive Technology exposes links on a page. 
In the previous version, if I tabbed through, I would find "Link", "Link", "Link". This is because the previous text (the context) does not get announced. The same would happen if I listed all the links on the page.
So, this PR puts the links as "wrapping" the context as well. They are visually distinct as links already, so the experience is now comparable for all users :)